refactor(runner): rename runner.rs to start.rs
Avoid confusing runner/src/runner.rs naming — the file implements
the start subcommand (poll loop, signal handling), so start.rs is
a clearer name.
